creat antismash downloader module
